Move packages
Does not make sense to have a utility package if essentially all classes under com.meti.lib function as utility classes
