+++ src/main/java/com/meti/lib/console/BufferedConsole.java
@@ -1,4 +1,4 @@
-package com.meti.lib.util.console;
+package com.meti.lib.console;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
