+++ src/main/java/com/meti/lib/convert/ClassConverter.java
@@ -1,4 +1,4 @@
-package com.meti.lib.util.convert;
+package com.meti.lib.convert;
 
 /**
  * @author SirMathhman
